// backend/src/common/enums/user-role.enum.ts
export enum UserRole {
  ADMIN = 'admin',
  SALES = 'sales',
  PLANNER = 'planner',
  INSPECTOR = 'inspector',
  TECHNICAL_MANAGER = 'technical_manager',
  FINANCE = 'finance',
  CUSTOMER_REP = 'customer_rep',
  EXECUTIVE = 'executive',
  CUSTOMER = 'customer', // Müşteri portali kullanıcısı
}